Include __init__.py file so that the addon can be installed from a .zip file
